feat: Move notification-related screens to notifications directory
This commit reorganizes the project structure by moving notification-related screens from the `lib/feature/home/ui/view` directory to `lib/feature/home/notifications/view` and `lib/feature/home/notifications/view/widgets`. This change improves code organization and maintainability by grouping related files together.

- Moved `PaymentDetailsScreen`, `InvoiceDetailsScreen`, and `NotificationsScreen` to the `lib/feature/home/notifications/view` directory.
- Moved `invoice_details_screen.dart` to `lib/feature/home/notifications/view/widgets`.
- Updated imports in `lib/core/router/route.dart` to reflect the new file structure.
- Adjusted padding in `NotificationsScreen` for better UI consistency.
