+++ lib/core/router/route.dart
@@ -3,9 +3,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:mzaodina_app/core/router/app_routes.dart';
 import 'package:mzaodina_app/feature/home/ui/view/home_details_screen.dart';
 import 'package:mzaodina_app/feature/home/ui/view/home_screen.dart';
-import 'package:mzaodina_app/feature/home/ui/view/invoice_details_screen.dart';
-import 'package:mzaodina_app/feature/home/ui/view/notifications_screen.dart';
-import 'package:mzaodina_app/feature/home/ui/view/payment_details_screen.dart';
+import 'package:mzaodina_app/feature/home/notifications/view/widgets/invoice_details_screen.dart';
+import 'package:mzaodina_app/feature/home/notifications/view/notifications_screen.dart';
+import 'package:mzaodina_app/feature/home/notifications/view/payment_details_screen.dart';
 import 'package:mzaodina_app/feature/nav_bar/view/nav_bar.dart';
 import 'package:mzaodina_app/feature/auth/ui/view-model/cubit/auth_cubit_cubit.dart';
 import 'package:mzaodina_app/feature/auth/ui/view/auth_screen.dart';
